fix: move PER related utility from common to domain/per
